+++ CHANGELOG.md
@@ -32,6 +32,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Harmonize setting names in `data-plane-transfer` (#1164)
 * Add Blob transfer Architectural Decision Record (#1259)
 * Add component tests coverage to the codecov coverage report (#1246)
+* Postgresql end to end test (#1278)
 
 #### Changed
 
@@ -43,6 +44,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Make data-plane public api controller asynchronous (#1228)
 * Provide In-mem implementations by default (#1130)
 * Changed Catalog config keys and switched from minutes to seconds
+* Uniform all the sql schema files as `docs/schema.sql` (#1278)
 * Clean FCC store before updating
 
 #### Removed